Aerial-Object-Detection-Bird-vs-Drone-main/requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.0
fsspec==2025.10.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
narwhals==2.12.0
networkx==3.4.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
polars==1.35.2
polars-runtime-32==1.35.2
protobuf==6.33.1
psutil==7.1.3
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
toml==0.10.2
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.233
ultralytics-thop==2.0.18
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.4
wrapt==2.0.1
